requirements.txt
scrapy
phonenumbers
python-slugify
twisted
klein
bs4
tldextract
langid